PR review comments

+++ src/cpp_generator.cpp
@@ -220,7 +220,7 @@ cpp_generator::function cpp_generator::generate_module(const module& m,
                         if(x < 0)
                             string_literal = "-__builtin_huge_val()";
                     }
-                    else if(std::isnan(x))
+                    else if(std::isnan(static_cast<double>(x)))
                         string_literal = "__builtin_nan(\"0\")";
                     else
                         string_literal = ins->get_literal().to_string();
